pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-auditor"
version = "0.1.0"
description = "Proactive code auditor agent for Unreal Engine C++ projects"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
code-auditor = "auditor.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/auditor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
